refactor: move directories
